Move serde under schemaregistry
