+++ schemaregistry/serializer/testhelpers.go
@@ -1,4 +1,4 @@
-package serde
+package serializer
 
 import (
 	"fmt"
